Refactor: add Local Audio and Video Control Buttons
