+++ src/components/CallView/shared/LocalAudioControlButton.vue
@@ -1,0 +1,162 @@
+<!--
+  - @copyright Copyright (c) 2023 Grigorii Shartsev <me@shgk.me>
+  -
+  - @author Grigorii Shartsev <me@shgk.me>
+  -
+  - @license GNU AGPL version 3 or any later version
+  -
+  - This program is free software: you can redistribute it and/or modify
+  - it under the terms of the GNU Affero General Public License as
+  - published by the Free Software Foundation, either version 3 of the
+  - License, or (at your option) any later version.
+  -
+  - This program is distributed in the hope that it will be useful,
+  - but WITHOUT ANY WARRANTY; without even the implied warranty of
+  - MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+  - GNU Affero General Public License for more details.
+  -
+  - You should have received a copy of the GNU Affero General Public License
+  - along with this program. If not, see <http://www.gnu.org/licenses/>.
+  -->
+
+<template>
+	<NcButton v-shortkey.once="disableKeyboardShortcuts ? null : ['m']"
+		v-tooltip="audioButtonTooltip"
+		:type="ncButtonType"
+		:aria-label="audioButtonAriaLabel"
+		:class="{ 'no-audio-available': !isAudioAllowed || !model.attributes.audioAvailable }"
+		@shortkey="toggleAudio"
+		@click.stop="toggleAudio">
+		<template #icon>
+			<VolumeIndicator :audio-preview-available="model.attributes.audioAvailable"
+				:audio-enabled="showMicrophoneOn"
+				:current-volume="model.attributes.currentVolume"
+				:volume-threshold="model.attributes.volumeThreshold"
+				:primary-color="color"
+				overlay-muted-color="#888888" />
+		</template>
+	</NcButton>
+</template>
+
+<script>
+import { emit } from '@nextcloud/event-bus'
+
+import { NcButton } from '@nextcloud/vue'
+
+import VolumeIndicator from '../../VolumeIndicator/VolumeIndicator.vue'
+
+import { PARTICIPANT } from '../../../constants.js'
+
+export default {
+	name: 'LocalAudioControlButton',
+
+	components: {
+		NcButton,
+		VolumeIndicator,
+	},
+
+	props: {
+		conversation: {
+			type: Object,
+			required: true,
+		},
+
+		model: {
+			type: Object,
+			required: true,
+		},
+
+		disableKeyboardShortcuts: {
+			type: Boolean,
+			default: OCP.Accessibility.disableKeyboardShortcuts(),
+		},
+
+		ncButtonType: {
+			type: String,
+			default: 'tertiary-no-background',
+		},
+
+		color: {
+			type: String,
+			default: 'currentColor',
+		},
+	},
+
+	computed: {
+		isAudioAllowed() {
+			return this.conversation.permissions & PARTICIPANT.PERMISSIONS.PUBLISH_AUDIO
+		},
+
+		showMicrophoneOn() {
+			return this.model.attributes.audioAvailable && this.model.attributes.audioEnabled
+		},
+
+		audioButtonTooltip() {
+			if (!this.isAudioAllowed) {
+				return t('spreed', 'You are not allowed to enable audio')
+			}
+
+			if (!this.model.attributes.audioAvailable) {
+				return {
+					content: t('spreed', 'No audio. Click to select device'),
+					show: false,
+				}
+			}
+
+			if (this.speakingWhileMutedNotification && !this.screenSharingMenuOpen) {
+				return {
+					content: this.speakingWhileMutedNotification,
+					show: true,
+				}
+			}
+
+			let content = ''
+			if (this.model.attributes.audioEnabled) {
+				content = this.disableKeyboardShortcuts
+					? t('spreed', 'Mute audio')
+					: t('spreed', 'Mute audio (M)')
+			} else {
+				content = this.disableKeyboardShortcuts
+					? t('spreed', 'Unmute audio')
+					: t('spreed', 'Unmute audio (M)')
+			}
+
+			return {
+				content,
+				show: false,
+			}
+		},
+
+		audioButtonAriaLabel() {
+			if (!this.model.attributes.audioAvailable) {
+				return t('spreed', 'No audio. Click to select device')
+			}
+
+			return this.model.attributes.audioEnabled
+				? t('spreed', 'Mute audio')
+				: t('spreed', 'Unmute audio')
+		},
+	},
+
+	methods: {
+		toggleAudio() {
+			if (!this.model.attributes.audioAvailable) {
+				emit('show-settings', {})
+				return
+			}
+
+			if (this.model.attributes.audioEnabled) {
+				this.model.disableAudio()
+			} else {
+				this.model.enableAudio()
+			}
+		},
+	},
+}
+</script>
+
+<style scoped>
+.no-audio-available {
+	opacity: .7;
+}
+</style>
