+++ src/components/CallView/shared/LocalVideoControlButton.vue
@@ -1,0 +1,164 @@
+<!--
+  - @copyright Copyright (c) 2023 Grigorii Shartsev <me@shgk.me>
+  -
+  - @author Grigorii Shartsev <me@shgk.me>
+  -
+  - @license GNU AGPL version 3 or any later version
+  -
+  - This program is free software: you can redistribute it and/or modify
+  - it under the terms of the GNU Affero General Public License as
+  - published by the Free Software Foundation, either version 3 of the
+  - License, or (at your option) any later version.
+  -
+  - This program is distributed in the hope that it will be useful,
+  - but WITHOUT ANY WARRANTY; without even the implied warranty of
+  - MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+  - GNU Affero General Public License for more details.
+  -
+  - You should have received a copy of the GNU Affero General Public License
+  - along with this program. If not, see <http://www.gnu.org/licenses/>.
+  -->
+
+<template>
+	<NcButton v-shortkey.once="disableKeyboardShortcuts ? null : ['v']"
+		v-tooltip="videoButtonTooltip"
+		:type="ncButtonType"
+		:aria-label="videoButtonAriaLabel"
+		:class="{ 'no-video-available': !isVideoAllowed || !model.attributes.videoAvailable }"
+		@shortkey="toggleVideo"
+		@click.stop="toggleVideo">
+		<template #icon>
+			<VideoIcon v-if="showVideoOn" :size="20" :fill-color="color" />
+			<VideoOff v-else :size="20" :fill-color="color" />
+		</template>
+	</NcButton>
+</template>
+
+<script>
+import VideoIcon from 'vue-material-design-icons/Video.vue'
+import VideoOff from 'vue-material-design-icons/VideoOff.vue'
+
+import { emit } from '@nextcloud/event-bus'
+
+import { NcButton } from '@nextcloud/vue'
+
+import { PARTICIPANT } from '../../../constants.js'
+
+export default {
+	name: 'LocalVideoControlButton',
+
+	components: {
+		NcButton,
+		VideoIcon,
+		VideoOff,
+	},
+
+	props: {
+		conversation: {
+			type: Object,
+			required: true,
+		},
+
+		model: {
+			type: Object,
+			required: true,
+		},
+
+		disableKeyboardShortcuts: {
+			type: Boolean,
+			default: OCP.Accessibility.disableKeyboardShortcuts(),
+		},
+
+		ncButtonType: {
+			type: String,
+			default: 'tertiary-no-background',
+		},
+
+		color: {
+			type: String,
+			default: 'currentColor',
+		},
+	},
+
+	computed: {
+		isVideoAllowed() {
+			return this.conversation.permissions & PARTICIPANT.PERMISSIONS.PUBLISH_VIDEO
+		},
+
+		showVideoOn() {
+			return this.model.attributes.videoAvailable && this.model.attributes.videoEnabled
+		},
+
+		videoButtonTooltip() {
+			if (!this.isVideoAllowed) {
+				return t('spreed', 'You are not allowed to enable video')
+			}
+
+			if (!this.model.attributes.videoAvailable) {
+				return t('spreed', 'No video. Click to select device')
+			}
+
+			if (this.model.attributes.videoEnabled) {
+				return this.disableKeyboardShortcuts
+					? t('spreed', 'Disable video')
+					: t('spreed', 'Disable video (V)')
+			}
+
+			if (!this.model.getWebRtc() || !this.model.getWebRtc().connection || this.model.getWebRtc().connection.getSendVideoIfAvailable()) {
+				return this.disableKeyboardShortcuts
+					? t('spreed', 'Enable video')
+					: t('spreed', 'Enable video (V)')
+			}
+
+			return this.disableKeyboardShortcuts
+				? t('spreed', 'Enable video - Your connection will be briefly interrupted when enabling the video for the first time')
+				: t('spreed', 'Enable video (V) - Your connection will be briefly interrupted when enabling the video for the first time')
+		},
+
+		videoButtonAriaLabel() {
+			if (!this.model.attributes.videoAvailable) {
+				return t('spreed', 'No video. Click to select device')
+			}
+
+			if (this.model.attributes.videoEnabled) {
+				return t('spreed', 'Disable video')
+			}
+
+			if (!this.model.getWebRtc() || !this.model.getWebRtc().connection || this.model.getWebRtc().connection.getSendVideoIfAvailable()) {
+				return t('spreed', 'Enable video')
+			}
+
+			return t('spreed', 'Enable video. Your connection will be briefly interrupted when enabling the video for the first time')
+		},
+	},
+
+	methods: {
+		toggleVideo() {
+			/**
+			 * Abort toggling the video if the 'v' key is lifted when pasting an
+			 * image in the new message form.
+			 */
+			if (document.getElementsByClassName('upload-editor').length !== 0) {
+				return
+			}
+
+			if (!this.model.attributes.videoAvailable) {
+				emit('show-settings', {})
+				return
+			}
+
+			if (this.model.attributes.videoEnabled) {
+				this.model.disableVideo()
+			} else {
+				this.model.enableVideo()
+			}
+		},
+	},
+}
+</script>
+
+<style scoped lang="scss">
+.no-video-available {
+	opacity: .7;
+}
+</style>
